packages/libs/utils/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "utils" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../../coverage/packages/libs/utils/html' --cov-report xml:'../../../coverage/packages/libs/utils/coverage.xml' --html='../../../reports/packages/libs/utils/unittests/html/index.html' --junitxml='../../../reports/packages/libs/utils/unittests/junit.xml'"

[tool.poetry]
name = "utils"
version = "1.0.0"
description = "This library includes GFW utils to reuse that code"
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "utils"

  [tool.poetry.dependencies]
  python = ">=3.9,<=3.12.2"
  pyyaml = "^6.0.1"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.0.2"
  flake8 = "6.0.0"
  pytest = "7.3.1"
  pytest-sugar = "0.9.7"
  pytest-cov = "4.1.0"
  pytest-html = "3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
